Define budgeting target icon as FileField
ImageField validator does not recognize svg file
as valid image file (as it relies on PIL)

https://code.djangoproject.com/ticket/14092

We changed it to FileField in order to allow
svg files to be uploaded.
Also, default value for target icon and Assignment image is set to null.
If null, default image should be set in frontend application